src/chapter-02/key-vault/requirements.txt
azure-identity
azure-keyvault-secrets
openai